* try to fix huge bug which delete .config folder on linux

+++ scripts/addons/Tila_ConfigManager/addon_manager/addon_manager.py
@@ -117,11 +117,12 @@ def disable_addon(addon_name):
     return True
 
 
-def file_acces_handler(func, path, exc_info):
-    LOG.debug("Handling Error for file ", path)
-    LOG.debug(exc_info)
+def file_acces_handler(func, path: Path, exc_info):
+
     # Check if file access issue
     if not os.access(path, os.W_OK):
+        LOG.debug(f"Handling Error for file {path}")
+        LOG.debug(exc_info)
         # Try to change the permision of file
         os.chmod(path, stat.S_IWUSR)
         # call the calling function again
@@ -301,6 +302,9 @@ class PathAM:
             if self.is_file:
                 os.remove(self.path)
             elif self.is_dir:
+                if not self.is_set:
+                    LOG.warning(f"skip removing invalid folder : {str(self.path)}")
+                    return
                 shutil.rmtree(self.path, onexc=file_acces_handler)
